完善请求查询的表格数据

+++ bee-apm-ui/bee-apm-ui-server/src/main/java/net/beeapm/ui/model/vo/TableVo.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class TableVo extends BaseVo{
     private List<Object> rows;
     private Integer pageNum;
-    private Integer pageTotal;
+    private Long pageTotal;
 
     public List<Object> getRows() {
         return rows;
@@ -23,11 +23,11 @@ public class TableVo extends BaseVo{
         this.pageNum = pageNum;
     }
 
-    public Integer getPageTotal() {
+    public Long getPageTotal() {
         return pageTotal;
     }
 
-    public void setPageTotal(Integer pageTotal) {
+    public void setPageTotal(Long pageTotal) {
         this.pageTotal = pageTotal;
     }
 }
